nba-mm/v2/requirements.txt
boto3>=1.34
pyarrow>=15.0
pandas>=2.0
polars>=1.0
duckdb>=1.0
python-dotenv>=1.0
